node posX, posY changed to double

+++ map-backend/src/main/java/pl/pwr/bdmap/dto/HistoricNodeDataDTO.java
@@ -4,8 +4,8 @@ import java.sql.Timestamp;
 
 public record HistoricNodeDataDTO(
         int id,
-        int posX,
-        int posY,
+        double posX,
+        double posY,
         Timestamp timestamp,
         int nodeId
 ) {}
